refactor(modals): extract position enum
