+++ src/UI/components/modals/modal.types.ts
@@ -1,0 +1,4 @@
+export enum PositionModal {
+    TOP,
+    BOTTOM
+}
